Rename xdp-redirect tool to xdp-bench
Rename the xdp-redirect tool to the more generic 'xdp-bench'. This is done
in preparation for adding support for benchmarking other XDP modes than
redirect.
