+++ Makefile
@@ -15,7 +15,7 @@ endif
 
 include version.mk
 
-UTILS := xdp-filter xdp-loader xdp-dump xdp-redirect xdp-monitor
+UTILS := xdp-filter xdp-loader xdp-dump xdp-bench xdp-monitor
 SUBDIRS := lib $(UTILS)
 .PHONY: check_submodule help clobber distclean clean install test libxdp $(SUBDIRS)
