+++ xdp-bench/Makefile
@@ -4,13 +4,13 @@ XDP_TARGETS := xdp_redirect_basic.bpf xdp_redirect_cpumap.bpf xdp_redirect_devma
 	       xdp_redirect_devmap_multi.bpf
 BPF_SKEL_TARGETS := $(XDP_TARGETS)
 
-TOOL_NAME := xdp-redirect
-MAN_PAGE := xdp-redirect.8
-USER_TARGETS := xdp-redirect
+TOOL_NAME := xdp-bench
+MAN_PAGE := xdp-bench.8
+USER_TARGETS := xdp-bench
 USER_LIBS     = -lm
 USER_EXTRA_C := xdp_redirect_basic.c xdp_redirect_cpumap.c xdp_redirect_devmap.c \
 		xdp_redirect_devmap_multi.c
-EXTRA_USER_DEPS := xdp_redirect.h
+EXTRA_USER_DEPS := xdp-bench.h
 
 LIB_DIR       = ../lib
