+++ xdp-bench/xdp-bench.8
@@ -1,7 +1,7 @@
-.TH "xdp-redirect" "8" "DECEMBER  7, 2022" "V1.2.2" "A simple XDP-powered packet redirect tool" 
+.TH "xdp-bench" "8" "DECEMBER  8, 2022" "V1.2.2" "A simple XDP benchmarking tool" 
 
 .SH "NAME"
-XDP-redirect \- a simple XDP-powered packet redirect tool
+XDP-bench \- a XDP simple benchmarking tool
 .SH "SYNOPSIS"
 .PP
 XDP-redirect is a packet redirecting utility powered by XDP. It is intended to
