+++ xdp-bench/xdp_redirect_cpumap.c
@@ -24,7 +24,7 @@
 
 #include "logging.h"
 
-#include "xdp_redirect.h"
+#include "xdp-bench.h"
 #include "xdp_sample.h"
 #include "xdp_redirect_cpumap.skel.h"
