+++ xdp-bench/xdp_redirect_devmap.c
@@ -20,7 +20,7 @@
 
 #include "logging.h"
 
-#include "xdp_redirect.h"
+#include "xdp-bench.h"
 #include "xdp_sample.h"
 #include "xdp_redirect_devmap.skel.h"
