+++ xdp-bench/xdp_redirect_devmap_multi.c
@@ -25,7 +25,7 @@
 #include "logging.h"
 
 #include "xdp_sample.h"
-#include "xdp_redirect.h"
+#include "xdp-bench.h"
 #include "xdp_redirect_devmap_multi.skel.h"
 
 static int ifaces[MAX_IFACE_NUM] = {};
