+++ xdp-monitor/xdp-monitor.8
@@ -1,4 +1,4 @@
-.TH "xdp-monitor" "8" "DECEMBER  7, 2022" "V1.2.2" "A simple XDP monitoring tool" 
+.TH "xdp-monitor" "8" "DECEMBER  8, 2022" "V1.2.2" "A simple XDP monitoring tool" 
 
 .SH "NAME"
 XDP-monitor \- a simple BPF-powered XDP monitoring
