--- xdp-redirect/.gitignore
@@ -1,1 +1,0 @@
-xdp-redirect
